+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TigerRouteSelectorTest.java
@@ -17,11 +17,13 @@
 package de.gematik.test.tiger.proxy;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import de.gematik.test.tiger.common.data.config.tigerproxy.ForwardProxyInfo;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerConfigurationRoute;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerProxyConfiguration;
 import java.util.List;
+import lombok.Cleanup;
 import lombok.SneakyThrows;
 import lombok.val;
 import org.junit.jupiter.api.Test;
@@ -45,14 +47,14 @@ class TigerRouteSelectorTest extends AbstractTigerProxyTest {
                         .build()))
             .build());
 
-    val response =
-        proxyRest.get("http://backend/foobar").asString();
+    val response = proxyRest.get("http://backend/foobar").asString();
     assertThat(response.getStatus()).isEqualTo(666);
   }
 
   @SneakyThrows
   @Test
   void twoDestinationsWithProxy() {
+    @Cleanup
     final TigerProxy forwardProxy =
         new TigerProxy(
             TigerProxyConfiguration.builder()
@@ -88,8 +90,76 @@ class TigerRouteSelectorTest extends AbstractTigerProxyTest {
                         .build()))
             .build());
 
-    val response =
-        proxyRest.get("http://backend/foobar").asString();
+    val response = proxyRest.get("http://backend/foobar").asString();
     assertThat(response.getStatus()).isEqualTo(666);
+  }
+
+  @Test
+  void twoUnreachableDestinations() {
+    assertThatExceptionOfType(IllegalStateException.class)
+        .isThrownBy( // NOSONAR
+            () -> {
+              spawnTigerProxyWith(
+                  TigerProxyConfiguration.builder()
+                      .proxyRoutes(
+                          List.of(
+                              TigerConfigurationRoute.builder()
+                                  .from("http://backend")
+                                  .to(
+                                      List.of(
+                                          "http://localhost:"
+                                              + TestSocketUtils.findAvailableTcpPort(),
+                                          "http://localhost:"
+                                              + TestSocketUtils.findAvailableTcpPort()))
+                                  .build()))
+                      .build());
+            })
+        .withMessageContaining("No reachable destination found among");
+  }
+
+  @Test
+  void twoUnreachableDestinationsWithProxy() {
+    @Cleanup
+    final TigerProxy forwardProxy =
+        new TigerProxy(
+            TigerProxyConfiguration.builder()
+                .proxyPort(TestSocketUtils.findAvailableTcpPort())
+                .adminPort(TestSocketUtils.findAvailableTcpPort())
+                .proxyRoutes(
+                    List.of(
+                        TigerConfigurationRoute.builder()
+                            .from("http://hostOnlyReachableViaProxy")
+                            .to("http://localhost:" + fakeBackendServerPort)
+                            .build(),
+                        TigerConfigurationRoute.builder()
+                            .from("/")
+                            .to("http://localhost:" + fakeBackendServerPort)
+                            .build()))
+                .build());
+
+    assertThatExceptionOfType(IllegalStateException.class)
+        .isThrownBy( // NOSONAR
+            () -> {
+              spawnTigerProxyWith(
+                  TigerProxyConfiguration.builder()
+                      .forwardToProxy(
+                          ForwardProxyInfo.builder()
+                              .hostname("localhost")
+                              .port(forwardProxy.getProxyPort())
+                              .build())
+                      .proxyRoutes(
+                          List.of(
+                              TigerConfigurationRoute.builder()
+                                  .from("http://backend")
+                                  .to(
+                                      List.of(
+                                          "http://localhost:"
+                                              + TestSocketUtils.findAvailableTcpPort(),
+                                          "http://localhost:"
+                                              + TestSocketUtils.findAvailableTcpPort()))
+                                  .build()))
+                      .build());
+            })
+        .withMessageContaining("No reachable destination found among");
   }
 }